Bugfix batch 1 to the parser
    After testing and other mucking about the first batch of patches is
ready to roll, which make the Java lexer/parser pair capable of reading
in a file containing a package def and a class with member vars and no
member functions. Member functions cause a stack overflow due to some
impropper recursion which has yet to be isolated. Imports still need to
be tested and interfaces are untouched. The main bug that this patch
fixes is getting the lexer and parser to agree on the base token, so
both files now generate and expect single-char tokens for all non-word
letters and word characters are dealt with in consecutive strings.